+++ src/@lib/equalities/shallowEquals.ts
@@ -1,8 +1,5 @@
 // Record 타입: 키-값 쌍을 가진 객체 타입을 정의할 때 사용
-export function shallowEquals<T>(
-  objA: T,
-  objB: T,
-): boolean {
+export function shallowEquals<T>(objA: T, objB: T): boolean {
   if (objA === objB) {
     // 같은 객체 참조 확인
     return true;
