+++ vite.config.ts
@@ -4,7 +4,8 @@ import react from "@vitejs/plugin-react-swc";
 
 export default mergeConfig(
   defineConfig({
-    base: process.env.NODE_ENV === "production" ? "/front_5th_chapter1-3/" : "/",
+    base:
+      process.env.NODE_ENV === "production" ? "/front_5th_chapter1-3/" : "/",
     plugins: [react()],
   }),
   defineTestConfig({
